Rename `GitHub` to `GitHubRead`
