+++ src/main.rs
@@ -4,7 +4,7 @@ mod team_api;
 mod utils;
 mod zulip;
 
-use crate::github::{GitHub, GitHubWrite, SyncGitHub};
+use crate::github::{GitHubRead, GitHubWrite, SyncGitHub};
 use crate::team_api::TeamApi;
 use anyhow::Context;
 use log::{error, info, warn};
@@ -81,13 +81,13 @@ fn app() -> anyhow::Result<()> {
         match service.as_str() {
             "github" => {
                 let token = get_env("GITHUB_TOKEN")?;
-                let github = GitHub::new(token.clone())?;
-                let sync = SyncGitHub::new(github, &team_api)?;
+                let gh_read = GitHubRead::new(token.clone())?;
+                let sync = SyncGitHub::new(gh_read, &team_api)?;
                 let diff = sync.diff_all()?;
                 info!("{}", diff);
                 if !only_print_plan {
-                    let github_write = GitHubWrite::new(token, dry_run)?;
-                    diff.apply(&github_write)?;
+                    let gh_write = GitHubWrite::new(token, dry_run)?;
+                    diff.apply(&gh_write)?;
                 }
             }
             "mailgun" => {
